Remove dataset doc embedding
Moved new document GET request functions out of experiments.js to experiments. Modified the dataset-rendering components to use the new document GET request functions. Finally, fixed a bug when using the hideGraph flag so that it no longer renders a graph tab — just the file table. ENCD-3455
