+++ src/encoded/types/dataset.py
@@ -323,15 +323,7 @@ class Annotation(FileSet, CalculatedBiosampleSlims, CalculatedBiosampleSynonyms)
         'targets',
         'files.dataset',
         'files.analysis_step_version.analysis_step',
-        'files.analysis_step_version.analysis_step.documents',
-        'files.analysis_step_version.analysis_step.documents.award',
-        'files.analysis_step_version.analysis_step.documents.lab',
-        'files.analysis_step_version.analysis_step.documents.submitted_by',
         'files.analysis_step_version.analysis_step.pipelines',
-        'files.analysis_step_version.analysis_step.pipelines.documents',
-        'files.analysis_step_version.analysis_step.pipelines.documents.award',
-        'files.analysis_step_version.analysis_step.pipelines.documents.lab',
-        'files.analysis_step_version.analysis_step.pipelines.documents.submitted_by',
         'files.analysis_step_version.analysis_step.versions',
         'files.analysis_step_version.analysis_step.versions.software_versions',
         'files.analysis_step_version.analysis_step.versions.software_versions.software',
@@ -504,9 +496,6 @@ class Series(Dataset, CalculatedSeriesAssay, CalculatedSeriesBiosample, Calculat
         'related_datasets.replicates.antibody',
         'related_datasets.replicates.antibody.targets',
         'related_datasets.replicates.library',
-        'related_datasets.replicates.library.documents.lab',
-        'related_datasets.replicates.library.documents.submitted_by',
-        'related_datasets.replicates.library.documents.award',
         'related_datasets.replicates.library.biosample.submitted_by',
         'related_datasets.replicates.library.biosample.source',
         'related_datasets.replicates.library.biosample.organism',
